+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetFormFactory/PaymentSheetFormFactory.swift
@@ -39,6 +39,7 @@ class PaymentSheetFormFactory {
     let currency: String?
     let amount: Int?
     let countryCode: String?
+    let cardBrandChoiceEligible: Bool
 
     var canSaveToLink: Bool {
         // For Link private beta, only save cards in ".none" mode: If there is no Customer object.
@@ -57,7 +58,8 @@ class PaymentSheetFormFactory {
         previousCustomerInput: IntentConfirmParams? = nil,
         addressSpecProvider: AddressSpecProvider = .shared,
         offerSaveToLinkWhenSupported: Bool = false,
-        linkAccount: PaymentSheetLinkAccount? = nil
+        linkAccount: PaymentSheetLinkAccount? = nil,
+        cardBrandChoiceEligible: Bool = false
     ) {
         func saveModeFor(merchantRequiresSave: Bool) -> SaveMode {
             let hasCustomer = configuration.hasCustomer
@@ -93,6 +95,7 @@ class PaymentSheetFormFactory {
                   addressSpecProvider: addressSpecProvider,
                   offerSaveToLinkWhenSupported: offerSaveToLinkWhenSupported,
                   linkAccount: linkAccount,
+                  cardBrandChoiceEligible: cardBrandChoiceEligible,
                   supportsLinkCard: intent.supportsLinkCard,
                   isPaymentIntent: intent.isPaymentIntent,
                   currency: intent.currency,
@@ -108,6 +111,7 @@ class PaymentSheetFormFactory {
         addressSpecProvider: AddressSpecProvider = .shared,
         offerSaveToLinkWhenSupported: Bool = false,
         linkAccount: PaymentSheetLinkAccount? = nil,
+        cardBrandChoiceEligible: Bool = false,
         supportsLinkCard: Bool,
         isPaymentIntent: Bool,
         currency: String?,
@@ -132,6 +136,7 @@ class PaymentSheetFormFactory {
         self.amount = amount
         self.countryCode = countryCode
         self.saveMode = saveMode
+        self.cardBrandChoiceEligible = cardBrandChoiceEligible
     }
 
     func make() -> PaymentMethodElement {
@@ -140,7 +145,7 @@ class PaymentSheetFormFactory {
         // We have two ways to create the form for a payment method
         // 1. Custom, one-off forms
         if paymentMethod == .card {
-            return makeCard()
+            return makeCard(cardBrandChoiceEligible: cardBrandChoiceEligible)
         } else if paymentMethod == .linkInstantDebit {
             return ConnectionsElement()
         } else if paymentMethod == .USBankAccount {
@@ -158,6 +163,8 @@ class PaymentSheetFormFactory {
             additionalElements = [makeRevolutPayMandate()]
         } else if paymentMethod.stpPaymentMethodType == .bancontact {
             return makeBancontact()
+        } else if paymentMethod.stpPaymentMethodType == .bacsDebit {
+            return makeBacsDebit()
         } else if paymentMethod.stpPaymentMethodType == .blik {
             return makeBLIK()
         } else if paymentMethod == .externalPayPal {
@@ -295,6 +302,38 @@ extension PaymentSheetFormFactory {
 
     func makeAUBECSMandate() -> StaticElement {
         return StaticElement(view: AUBECSLegalTermsView(configuration: configuration))
+    }
+
+    func makeSortCode() -> PaymentMethodElementWrapper<TextFieldElement> {
+        let defaultValue = previousCustomerInput?.paymentMethodParams.bacsDebit?.sortCode
+        let element = TextFieldElement.Account.makeSortCode(defaultValue: defaultValue, theme: theme)
+        return PaymentMethodElementWrapper(element) { textField, params in
+            let sortCodeText = BSBNumber(number: textField.text).bsbNumberText()
+            params.paymentMethodParams.nonnil_bacsDebit.sortCode = sortCodeText
+            return params
+        }
+    }
+
+    func makeBacsAccountNumber() -> PaymentMethodElementWrapper<TextFieldElement> {
+        let defaultValue = previousCustomerInput?.paymentMethodParams.bacsDebit?.accountNumber
+        let element = TextFieldElement.Account.makeBacsAccountNumber(defaultValue: defaultValue, theme: theme)
+        return PaymentMethodElementWrapper(element) { textField, params in
+            params.paymentMethodParams.nonnil_bacsDebit.accountNumber = textField.text
+            return params
+        }
+    }
+
+    func makeBacsMandate() -> PaymentMethodElementWrapper<CheckboxElement> {
+        let mandateText = String(format: String.Localized.bacs_mandate_text, configuration.merchantDisplayName)
+        let element = CheckboxElement(
+            theme: configuration.appearance.asElementsTheme,
+            label: mandateText,
+            isSelectedByDefault: false
+        )
+        return PaymentMethodElementWrapper(element) { checkbox, params in
+            // Only return params if the mandate has been accepted
+            return checkbox.isSelected ? params : nil
+        }
     }
 
     func makeSepaMandate() -> PaymentMethodElement {
@@ -449,6 +488,28 @@ extension PaymentSheetFormFactory {
         let addressSection: Element? = makeBillingAddressSectionIfNecessary(requiredByPaymentMethod: false)
         let mandate: Element? = saveMode == .merchantRequired ? makeSepaMandate() : nil // Note: We show a SEPA mandate b/c iDEAL saves bank details as a SEPA Direct Debit Payment Method
         let elements: [Element?] = [contactSection, addressSection, mandate]
+        return FormElement(
+            autoSectioningElements: elements.compactMap { $0 },
+            theme: theme
+        )
+    }
+
+    func makeBacsDebit() -> PaymentMethodElement {
+        let contactSection: Element? = makeContactInformationSection(
+            nameRequiredByPaymentMethod: true,
+            emailRequiredByPaymentMethod: true,
+            phoneRequiredByPaymentMethod: false
+        )
+        let addressSection: Element? = makeBillingAddressSectionIfNecessary(requiredByPaymentMethod: true)
+        let sortCodeField = makeSortCode()
+        let accountNumberField = makeBacsAccountNumber()
+        let mandate = makeBacsMandate()
+        let bacsAccountSection = SectionElement(
+            title: String.Localized.bank_account_sentence_case,
+            elements: [sortCodeField, accountNumberField],
+            theme: theme
+        )
+        let elements: [Element?] = [contactSection, bacsAccountSection, addressSection, mandate]
         return FormElement(
             autoSectioningElements: elements.compactMap { $0 },
             theme: theme