+++ StripeUICore/StripeUICore/Source/Elements/DropdownFieldElement.swift
@@ -20,18 +20,27 @@ import UIKit
     public typealias DidUpdateSelectedIndex = (Int) -> Void
 
     public struct DropdownItem {
-        public init(pickerDisplayName: String, labelDisplayName: String, accessibilityValue: String, rawData: String) {
+        public init(pickerDisplayName: NSAttributedString, labelDisplayName: NSAttributedString, accessibilityValue: String, rawData: String, isPlaceholder: Bool = false) {
             self.pickerDisplayName = pickerDisplayName
             self.labelDisplayName = labelDisplayName
             self.accessibilityValue = accessibilityValue
+            self.isPlaceholder = isPlaceholder
             self.rawData = rawData
         }
 
+        public init(pickerDisplayName: String, labelDisplayName: String, accessibilityValue: String, rawData: String, isPlaceholder: Bool = false) {
+            self = .init(pickerDisplayName: NSAttributedString(string: pickerDisplayName),
+                         labelDisplayName: NSAttributedString(string: labelDisplayName),
+                         accessibilityValue: accessibilityValue,
+                         rawData: rawData,
+                         isPlaceholder: isPlaceholder)
+        }
+
         /// Item label displayed in the picker
-        public let pickerDisplayName: String
+        public let pickerDisplayName: NSAttributedString
 
         /// Item label displayed in inline label when item has been selected
-        public let labelDisplayName: String
+        public let labelDisplayName: NSAttributedString
 
         /// Accessibility value to use when this is in the inline label
         public let accessibilityValue: String
@@ -40,11 +49,14 @@ import UIKit
         /// e.g., A country dropdown item might display "United States" but its `rawData` is "US".
         /// This is ignored by `DropdownFieldElement`, and is intended as a convenience to be used in conjunction with `selectedItem`
         public let rawData: String
+
+        /// If true, this item will be styled with greyed out secondary text
+        public let isPlaceholder: Bool
     }
 
     // MARK: - Public properties
     weak public var delegate: ElementDelegate?
-    public let items: [DropdownItem]
+    public private(set) var items: [DropdownItem]
     public var selectedItem: DropdownItem {
         return items[selectedIndex]
     }
@@ -54,6 +66,9 @@ import UIKit
         }
     }
     public var didUpdate: DidUpdateSelectedIndex?
+    public let theme: ElementsUITheme
+    public let hasPadding: Bool
+
     /// A label displayed in the dropdown field UI e.g. "Country or region" for a country dropdown
     public let label: String?
 #if targetEnvironment(macCatalyst)
@@ -66,7 +81,7 @@ import UIKit
         if #available(macCatalyst 14.0, *) {
             let menu = UIMenu(children:
                 items.enumerated().map { (index, item) in
-                    UIAction(title: item.pickerDisplayName, identifier: .init(rawValue: String(index)), handler: action)
+                    UIAction(title: item.pickerDisplayName.string, identifier: .init(rawValue: String(index)), handler: action)
                 }
             )
             button.menu = menu
@@ -92,7 +107,8 @@ import UIKit
             shouldShowChevron: disableDropdownWithSingleElement ? items.count != 1 : true,
             pickerView: pickerView,
             delegate: self,
-            theme: theme
+            theme: theme,
+            hasPadding: hasPadding
         )
         if disableDropdownWithSingleElement && items.count == 1 {
             pickerFieldView.isUserInteractionEnabled = false
@@ -101,27 +117,27 @@ import UIKit
     }()
 
     // MARK: - Private properties
-    private let theme: ElementsUITheme
     private var previouslySelectedIndex: Int
     private let disableDropdownWithSingleElement: Bool
 
     /**
      - Parameters:
-       - items: Items to populate this dropdown with.
-       - defaultIndex: Defaults the dropdown to the item with the corresponding index.
-       - label: Label for the dropdown
-       - didUpdate: Called when the user has finished selecting a new item.
-
+     - items: Items to populate this dropdown with.
+     - defaultIndex: Defaults the dropdown to the item with the corresponding index.
+     - label: Label for the dropdown
+     - didUpdate: Called when the user has finished selecting a new item.
+     
      - Note:
-       - Items must contain at least one item.
-       - If `defaultIndex` is outside of the bounds of the `items` array, then a default of `0` is used.
-       - `didUpdate` is not called if the user does not change their input before hitting "Done"
+     - Items must contain at least one item.
+     - If `defaultIndex` is outside of the bounds of the `items` array, then a default of `0` is used.
+     - `didUpdate` is not called if the user does not change their input before hitting "Done"
      */
     public init(
         items: [DropdownItem],
         defaultIndex: Int = 0,
         label: String?,
         theme: ElementsUITheme = .default,
+        hasPadding: Bool = true,
         disableDropdownWithSingleElement: Bool = false,
         didUpdate: DidUpdateSelectedIndex? = nil
     ) {
@@ -132,6 +148,7 @@ import UIKit
         self.items = items
         self.disableDropdownWithSingleElement = disableDropdownWithSingleElement
         self.didUpdate = didUpdate
+        self.hasPadding = hasPadding
 
         // Default to defaultIndex, if in bounds
         if defaultIndex < 0 || defaultIndex >= items.count {
@@ -147,9 +164,18 @@ import UIKit
         }
     }
 
-    public func select(index: Int) {
+    public func select(index: Int, shouldAutoAdvance: Bool = true) {
         selectedIndex = index
-        didFinish(pickerFieldView)
+        didFinish(pickerFieldView, shouldAutoAdvance: shouldAutoAdvance)
+    }
+
+    public func update(items: [DropdownItem]) {
+        assert(!items.isEmpty, "`items` must contain at least one item")
+        // Try to re-select the same item afer updating, if not possible default to the first item in the list
+        let newSelectedIndex = items.firstIndex(where: { $0.rawData == self.items[selectedIndex].rawData }) ?? 0
+
+        self.items = items
+        self.select(index: newSelectedIndex, shouldAutoAdvance: false)
     }
 }
 
@@ -164,6 +190,7 @@ private extension DropdownFieldElement {
         }
         #else
         if pickerView.selectedRow(inComponent: 0) != selectedIndex {
+            pickerView.reloadComponent(0)
             pickerView.selectRow(selectedIndex, inComponent: 0, animated: false)
         }
         #endif
@@ -189,8 +216,16 @@ extension DropdownFieldElement: Element {
 // MARK: UIPickerViewDelegate
 
 extension DropdownFieldElement: UIPickerViewDelegate {
-    public func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
-        return items[row].pickerDisplayName
+
+    public func pickerView(_ pickerView: UIPickerView, attributedTitleForRow row: Int, forComponent component: Int) -> NSAttributedString? {
+        let item = items[row]
+
+        guard item.isPlaceholder else { return item.pickerDisplayName }
+
+        // If this item is marked as a placeholder, apply placeholder text color
+        let attributes: [NSAttributedString.Key: Any] = [.foregroundColor: theme.colors.placeholderText]
+        let placeholderString = NSAttributedString(string: item.pickerDisplayName.string, attributes: attributes)
+        return placeholderString
     }
 
     public func pickerView(_ pickerView: UIPickerView, didSelectRow row: Int, inComponent component: Int) {
@@ -215,11 +250,22 @@ extension DropdownFieldElement: PickerFieldViewDelegate {
         // No-op
     }
 
-    func didFinish(_ pickerFieldView: PickerFieldView) {
+    func didFinish(_ pickerFieldView: PickerFieldView, shouldAutoAdvance: Bool) {
         if previouslySelectedIndex != selectedIndex {
             didUpdate?(selectedIndex)
         }
         previouslySelectedIndex = selectedIndex
-        delegate?.continueToNextField(element: self)
+
+        if shouldAutoAdvance {
+            DispatchQueue.main.async { [weak self] in
+                guard let self = self else { return }
+                self.delegate?.continueToNextField(element: self)
+            }
+        }
+    }
+
+    func didCancel(_ pickerFieldView: PickerFieldView) {
+        // Reset to previously selected index when canceling
+        selectedIndex = previouslySelectedIndex
     }
 }